backend/requirements.txt
Django==2.2.19
pytz==2022.2.1
sqlparse==0.4.2
djangorestframework==3.13.1
Pillow==9.2.0